+++ mayan/apps/document_states/models/workflow_state_escalation_model_mixins.py
@@ -1,12 +1,9 @@
 import hashlib
 
-from django.conf import settings
 from django.core import serializers
 from django.utils.translation import gettext_lazy as _
 
-from ..literals import (
-    ERROR_LOG_DOMAIN_NAME, GRAPHVIZ_SYMBOL_CONDITIONAL, GRAPHVIZ_SYMBOL_TIME
-)
+from ..literals import GRAPHVIZ_SYMBOL_CONDITIONAL, GRAPHVIZ_SYMBOL_TIME
 
 
 class WorkflowStateEscalationBusinessLogicMixin:
@@ -33,29 +30,6 @@ class WorkflowStateEscalationBusinessLogicMixin:
                 'tail_name': self.state.get_graph_id()
             }
             diagram.edge(**edge_kwargs)
-
-    def execute(self, context, workflow_instance):
-        condition_context = {'workflow_instance': workflow_instance}
-        condition_result = self.evaluate_condition(context=condition_context)
-
-        if condition_result:
-            try:
-                self.get_class_instance().execute(context=context)
-            except Exception as exception:
-                self.error_log.create(
-                    domain_name=ERROR_LOG_DOMAIN_NAME,
-                    text='{}; {}'.format(
-                        exception.__class__.__name__, exception
-                    )
-                )
-
-                if settings.DEBUG or settings.TESTING:
-                    raise
-            else:
-                queryset_error_logs = self.error_log.filter(
-                    domain_name=ERROR_LOG_DOMAIN_NAME
-                )
-                queryset_error_logs.delete()
 
     def get_comment(self):
         return self.comment or _(message='Workflow escalation.')