+++ mayan/apps/document_states/models/workflow_state_model_mixins.py
@@ -1,6 +1,7 @@
 import hashlib
 
 from django.apps import apps
+from django.conf import settings
 from django.core import serializers
 from django.db.models import F, Max, Q
 from django.utils.translation import gettext_lazy as _
@@ -8,13 +9,55 @@ from django.utils.translation import gettext_lazy as _
 from mayan.apps.documents.models.document_models import Document
 
 from ..literals import (
-    GRAPHVIZ_COLOR_STATE_FILL, GRAPHVIZ_ID_STATE, GRAPHVIZ_SHAPE_CIRCLE,
-    GRAPHVIZ_SHAPE_DOUBLECIRCLE, GRAPHVIZ_STYLE_FILLED,
+    ERROR_LOG_DOMAIN_NAME, GRAPHVIZ_COLOR_STATE_FILL, GRAPHVIZ_ID_STATE,
+    GRAPHVIZ_SHAPE_CIRCLE, GRAPHVIZ_SHAPE_DOUBLECIRCLE, GRAPHVIZ_STYLE_FILLED,
     WORKFLOW_ACTION_ON_ENTRY, WORKFLOW_ACTION_ON_EXIT
 )
 
 
 class WorkflowStateBusinessLogicMixin:
+    def do_active_set(self, workflow_instance):
+        queryset = self.entry_actions.filter(enabled=True)
+        self.do_queryset_actions_execute(
+            queryset=queryset, workflow_instance=workflow_instance
+        )
+
+    def do_active_unset(self, workflow_instance):
+        queryset = self.exit_actions.filter(enabled=True)
+        self.do_queryset_actions_execute(
+            queryset=queryset, workflow_instance=workflow_instance
+        )
+
+    def do_queryset_actions_execute(self, queryset, workflow_instance):
+        for action in queryset:
+            context = workflow_instance.get_context()
+            context.update(
+                {'action': action}
+            )
+
+            try:
+                action.execute(
+                    context=context, workflow_instance=workflow_instance
+                )
+            except Exception as exception:
+                queryset_error_logs = workflow_instance.document.error_log
+                queryset_error_logs.create(
+                    domain_name=ERROR_LOG_DOMAIN_NAME,
+                    text='{}; {}'.format(
+                        exception.__class__.__name__, exception
+                    )
+                )
+
+                if settings.DEBUG or settings.TESTING:
+                    raise
+
+                break
+            else:
+                queryset_error_logs = workflow_instance.document.error_log.filter(
+                    domain_name=ERROR_LOG_DOMAIN_NAME
+                )
+                queryset_error_logs.delete()
+
     def do_diagram_generate(self, diagram):
         is_edge_state = self.initial or not self.destination_transitions.exists() or not self.origin_transitions.exists()